test(cli): extract town focus play tests
Move the town focus and town-project review command tests into a focused play suite. The new file owns the city-command/city-operation validation and send fake state for CHANGE_GROWTH_MODE and CONSIDER_TOWN_PROJECT.

Keep the monolith responsible for read-side town-focus, ready-city, notification, and compact-priority projection fixtures. Wire the focused suite into test:cli:play so the extracted coverage remains canonical.

Proof: git diff --check; bun run --cwd packages/cli test -- game.play.town-focus.test.ts; bun run --cwd packages/cli test -- game.play.test.ts -t "town focus|town project|ready-city|compact priorities|notification queue"; bun run check:cli; bun run test:cli:play. This is test ownership modularization only; it changes no runtime behavior, direct-control behavior, live-game flow, gameplay send, play-thread workflow, Studio/oRPC transport, or relationship/city-state/suzerain evidence policy.
